Revert "Revert "Remove `include/grpcpp/impl/codegen/async_unary_call.h` (#30945)" (#30950)"
This reverts commit 0c454fa8758363f99d0b29e849250758191adc4a.
